requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pandas
langchain
langchain-google-genai
chromadb
python-dotenv
streamlit